move not so interesting stuff out of FailGood.kt
