+++ failgood/jvm/src/failgood/ExecutionListener.kt
@@ -1,6 +1,6 @@
 package failgood
 
-interface ExecutionListener {
+internal interface ExecutionListener {
     suspend fun testStarted(testDescription: TestDescription)
     suspend fun testFinished(testPlusResult: TestPlusResult)
     suspend fun testEvent(testDescription: TestDescription, type: String, payload: String)
