+++ failgood/jvm/src/failgood/Suite.kt
@@ -87,7 +87,7 @@ class Suite(val contextProviders: Collection<ContextProvider>) {
     }
 }
 
-object NullExecutionListener : ExecutionListener {
+internal object NullExecutionListener : ExecutionListener {
     override suspend fun testStarted(testDescription: TestDescription) {}
     override suspend fun testFinished(testPlusResult: TestPlusResult) {}
     override suspend fun testEvent(testDescription: TestDescription, type: String, payload: String) {}
