+++ failgood/jvm/src/failgood/internal/ExceptionPrettyPrinter.kt
@@ -1,7 +1,5 @@
 package failgood.internal
 
-import failgood.SourceInfo
-
 internal class ExceptionPrettyPrinter(private val throwable: Throwable, sourceInfo: SourceInfo? = null) {
     val stackTrace = run {
         val onlyElementsWithLineNumber = throwable.stackTrace
