+++ failgood/jvm/test/failgood/SuiteTest.kt
@@ -1,6 +1,7 @@
 package failgood
 
 import failgood.internal.FailedRootContext
+import failgood.internal.RootContext
 import failgood.mock.mock
 import kotlinx.coroutines.*
 import strikt.api.expectThat
