+++ failgood/jvm/test/failgood/ThousandTestsTest.kt
@@ -1,5 +1,6 @@
 package failgood
 
+import failgood.internal.RootContext
 import strikt.api.expectThat
 import strikt.assertions.isTrue
