+++ failgood/jvm/test/failgood/internal/TestResultFixtures.kt
@@ -2,7 +2,6 @@ package failgood.internal
 
 import failgood.Context
 import failgood.Failure
-import failgood.SourceInfo
 import failgood.Success
 import failgood.TestDescription
 import failgood.TestPlusResult
