+++ failgood/jvm/test/failgood/junit/CreateResponseTest.kt
@@ -1,7 +1,6 @@
 package failgood.junit
 
 import failgood.Context
-import failgood.SourceInfo
 import failgood.Success
 import failgood.Test
 import failgood.TestDescription
@@ -9,6 +8,7 @@ import failgood.TestPlusResult
 import failgood.describe
 import failgood.internal.ContextInfo
 import failgood.internal.FailedRootContext
+import failgood.internal.SourceInfo
 import failgood.internal.SuiteExecutionContext
 import kotlinx.coroutines.CompletableDeferred
 import org.junit.platform.engine.TestDescriptor
